Removed JGoodies dependency. Added date panel size to DatePickerSettings..
This was a major overhaul of the project. All dependencies on JGoodies
were removed. All panels now use GridBagLayout, or grid layout for some
panels. The size of the date labels panel (and indirectly the date
picker calendar) can now be set in the DatePickerSettings. The
background color of the weekday labels was added to DatePickerSettings.