Major update for version  2.4.1. See commit description.
Rewrote the Demo application. This included adding multiple translations
to the demo, and adding the new date picker features to the demo
Integrated all component listeners used by the CustomPopup class,  to be
internal to the class.
Added colors and the popup border to the date picker settings.
Changed version number to 2.4.1.