Adjusted dependencies to work with maven central.
All needed classes for JGoodies Forms 1.7.2 and it's dependencies are
now included directly in the project, under the package name
"com.privatejgoodies". Maven shade is no longer being used to include or
rename these dependencies. This dependency is no longer required or
listed in the main pom file. The list of classes that was being used by
the project was acquired by extracting older versions of the project jar
that had been minimized with shade. A zipped working version of the
jgoodies form source (with dependencies, and -all- original files), was
stored in the "zExtraFiles" folder of the repository, in case more
features of the library are needed later that require additional
classes.