Fixed a null pointer bug, and added the independent CalendarPanel component.
Fixed a bug where the date picker was causing a null pointer exception
on some operating systems. Changed the CalendarPanel class so that it
can be (optionally) used independently of the date picker. Added a
"basic demo" to the library. Added an "independent calendar panel demo"
to the library. Changed name of the main demo class to "FullDemo.java".