Adapted the library for the Java 1.6 Backport.
Adapted the library to make all source code compatible with Java 1.6.
Added the JSR 310 Backport library as an optional dependency in the POM
file.
Added instructions to the POM file for switching between Java 1.8, and
Java 1.6 (Backport) release versions.
Fixed the default Turkish language date format to remove the name of the
weekday.
Added a button to the demo to show the current Java versions.
Performed a recursive styling format on all files.