Adapted the DatePickerSettings to allow dynamic settings changes in the independent CalendarPanel.
Adapted the DatePickerSettings to allow dynamic settings changes in the
independent CalendarPanel. All DatePickerSettings now use accessor
functions for public access. Modified the CalendarPanel to detect and
handle changes in the settings after an independent CalendarPanel is
constructed. Changed the internal workings of the calendar panel layout
for greater control of the edge spacing when the calendar panel is
resized.