Updated docs in the format action.
We only use a 'push' trigger, and not a 'pull_request' trigger. 

Reason: From testing, it appears that when a 'pull_request' trigger is used, it causes the format action to fail during any pull request that originates from any forked repository. 

This change fixes issues discussed in the pull requests listed here. Pull requests, #152, #156, #158, #159, #160, #161, #163, #164.