Added FlatLAF to the demo classes.
Added the 'FlatLAF' look and feel to the demo classes, because the default 'Metal' look and feel looks outdated. This required adding the FlatLAF library as a maven dependency. That dependency should not be a problem for parent projects, because parent projects can override the dependency as desired. (Including the library version.)